rest/requirements.txt
cmake
face_recognition
numpy
uvicorn
python-multipart
opencv-python
textblob
email-validator
mysql-connector-python
